added Enity class to OpenSteer shutdown and loop review

+++ Source/EduNetGames/Tutorial_03/NetBoidFactory.h
@@ -17,7 +17,7 @@ public:
 private:
 
 	NetBoidReplicaManager* m_pkReplicaManager;
-	DataStructures::Map<unsigned int, RakNet::Replica3* > m_uidMap;
+	DataStructures::Map<size_t, RakNet::Replica3* > m_uidMap;
 };
 
 // ----------------------------------------------------------------------------
