+++ Source/EduNetGames/Tutorial_04/plugins/NetPedestrianReplicaConnection.h
@@ -56,7 +56,7 @@ public:
 private:
 
 	NetPedestrianReplicaManager* m_pkReplicaManager;
-	DataStructures::Map<unsigned int, RakNet::Replica3* > m_uidMap;
+	DataStructures::Map<size_t, RakNet::Replica3* > m_uidMap;
 };
 
 //-----------------------------------------------------------------------------
